+++ example/errors.py
@@ -1,5 +1,0 @@
-"""Custom exceptions."""
-
-
-class DownTimeError(Exception):
-    """Custom exception to indicate ESI is in daily downtime."""