+++ example/tests/__init__.py
@@ -9,8 +9,7 @@ from django.core.handlers.wsgi import WSGIRequest
 from django.test import RequestFactory, TestCase
 
 # AA Example
-from example.tests.testdata.integrations.allianceauth import load_allianceauth
-from example.tests.testdata.utils import create_user_from_evecharacter
+from example.tests.testdata.example import UserMainFactory
 
 
 class SocketAccessError(Exception):
@@ -49,19 +48,8 @@ class ExampleTestCase(NoSocketsTestCase):
     """
     Preloaded Testcase class for Example tests without Network access.
 
-    Pre-Load:
-        * Alliance Auth Characters, Corporation, Alliance Data
-        * Taken User IDs: 1001
-
     Available Request Factory:
         `self.factory`
-
-    Available test users:
-        * `user` User with standard Example access.
-            * 'example.basic_access' Permission
-            * Character ID 1001
-            * Corporation ID 2001
-            * Alliance ID 3001
 
     Example:
         .. code-block:: python
@@ -74,17 +62,17 @@ class ExampleTestCase(NoSocketsTestCase):
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
-        # Initialize Alliance Auth test data
-        load_allianceauth()
 
         # Request Factory
         cls.factory = RequestFactory()
 
-        # User with Standard Access - Corporation 2001
-        cls.user, cls.user_character = create_user_from_evecharacter(
-            character_id=1001,
-            permissions=["example.basic_access"],
-        )
+        # User with Standard Access
+        cls.user = UserMainFactory()
+
+        # User with Superuser Access
+        cls.superuser = UserMainFactory()
+        cls.superuser.is_superuser = True
+        cls.superuser.save()
 
     def _middleware_process_request(self, request: WSGIRequest):
         """Helper method to process middleware for a request."""