+++ example/tests/testdata/utils.py
@@ -7,7 +7,6 @@ import string
 from django.contrib.auth.models import User
 
 # Alliance Auth
-from allianceauth.authentication.backends import StateBackend
 from allianceauth.authentication.models import CharacterOwnership
 from allianceauth.eveonline.models import EveCharacter
 from allianceauth.tests.auth_utils import AuthUtils
@@ -181,31 +180,6 @@ def add_character_to_user(
         AuthUtils.connect_signals()
 
     return CharacterOwnership.objects.get(user=user, character=character)
-
-
-def create_user_from_evecharacter(
-    character_id: int,
-    permissions: list[str] | None = None,
-    scopes: list[str] | None = None,
-) -> tuple[User, CharacterOwnership]:
-    """Create new allianceauth user from EveCharacter object.
-
-    Args:
-        character_id (int): ID of eve character
-        permissions (list[str] | None): list of permission names, e.g. `"my_app.my_permission"`
-        scopes (list[str] | None): list of scope names
-    Returns:
-        tuple(User, CharacterOwnership): Created Alliance Auth User and CharacterOwnership
-    """
-    auth_character = EveCharacter.objects.get(character_id=character_id)
-    user = AuthUtils.create_user(auth_character.character_name.replace(" ", "_"))
-    character_ownership = add_character_to_user(
-        user, auth_character, is_main=True, scopes=scopes
-    )
-    if permissions:
-        for permission_name in permissions:
-            user = AuthUtils.add_permission_to_user_by_name(permission_name, user)
-    return user, character_ownership
 
 
 def add_permission_to_user(